requirements.txt
# Flask==2.2.3
# openai==0.27.8
# python-dotenv==1.0.0
# requests==2.28.2
Flask==2.2.3
Werkzeug==2.2.3
openai==0.27.8
python-dotenv==1.0.0
requests==2.28.2